DS18B20原先的没错

+++ 13-1 DS18B20温度读取/DS18B20.h
@@ -1,6 +1,8 @@
 #ifndef __DS18B20_H__
 #define __DS18B20_H__
 
-int DS18B20_GetTemp(void);
+// void DS18B20_Convert(void);
+float DS18B20_GetTemp(void);
+// float DS18B20_ReadT(void);
 
 #endif
